Move punycode in its own file. Update LICENSE accordingly. Enable back 2 test cases and update urlParse to make sure we skip only invalid chars in the ASCII range.
